referral_system/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.11
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
sqlparse==0.4.4
typing-extensions==4.11.0
tzdata==2024.1
utils==1.0.2
